+++ app/server/createImageFromDallEAPI/createImageFromDallEAPI.ts
@@ -1,15 +1,14 @@
 import { Configuration, OpenAIApi } from "openai";
 import { setTimeout } from "timers/promises";
-// import { MOCK_BASE64_IMAGE } from "~/__mocks__/base64Image";
 import { addBase64EncodedImageToAWS, addNewImageToDB } from "~/server";
 import { getS3BucketURL } from "~/utils";
+import { getMockDataResponse } from "./utils";
 
 const configuration = new Configuration({
   apiKey: process.env.DALLE_API_KEY,
 });
 const openai = new OpenAIApi(configuration);
 
-const MOCK_IMAGE_ID = "cliid9qad0001r2q9pscacuj0";
 const DEFAULT_NUMBER_OF_IMAGES_CREATED = 1;
 const IMAGE_SIZE = "1024x1024";
 const DEFAULT_AI_IMAGE_LANGUAGE_MODEL = "dall-e";
@@ -26,7 +25,7 @@ const DEFAULT_PAYLOAD = {
  * @description
  * This function makes a request to Open AI's Dall-E API to fetch images generated using the prompt
  */
-const generateImages = async (
+const createDallEImages = async (
   prompt: string,
   numberOfImages = DEFAULT_NUMBER_OF_IMAGES_CREATED
 ) => {
@@ -65,29 +64,14 @@ export const createImageFromDallEAPI = async (
       console.log(
         "\x1b[33m ⚠️ Warning – Using Mock Data ************************* \x1b[0m"
       );
-      const imageURL = getS3BucketURL(MOCK_IMAGE_ID);
-      const mockDallEImage = {
-        id: MOCK_IMAGE_ID,
-        prompt: "using mock data",
-        userId: "testUser123",
-        createdAt: "2023-06-26 03:17:19",
-        user: {
-          userId: "123456789",
-          username: "testUser123",
-        },
-        url: imageURL,
-        comments: [],
-      };
-      const mockArrayOfDallEGeneratedImages = new Array(numberOfImages).fill(
-        mockDallEImage
-      );
+      const mockData = getMockDataResponse(numberOfImages);
       await setTimeout(THREE_SECONDS_IN_MS);
 
-      return { images: mockArrayOfDallEGeneratedImages };
+      return { images: mockData };
     }
 
     // Generate Images
-    const imagesImages = await generateImages(prompt, numberOfImages);
+    const imagesImages = await createDallEImages(prompt, numberOfImages);
 
     const formattedImageData = await Promise.all(
       imagesImages.map(async (imageImage) => {
